# Replace slow AccessTools.TypeByName with a faster version