This replaces `AccessTools.TypeByName(name)` in `CreateAssemblyCopiesWithInternalsVisibleTo` with a custom version. `AccessTools` does a complete enumeration of every type in every assembly. We can instead just do `assembly.GetType(name)` for each assembly. I've also added a cache, because the search cost is linear in the number of assemblies.

I was seeing very erratic FSR performance on my machine - 400ms if we're feeling lucky, but spiking to multiple seconds - occasionally 10s+! I don't fully understand why it would be so erratic, but `AccessTools.TypeByName` was the major culprit. It was sometimes taking 900ms per call! This new version takes ~0.5ms per call.

I'm 90% sure that this is functionally equivalent to what was there before, but would appreciate a second pair of eyes to review that.