@handzlikchris 2023-06-25
Thanks @SamPruden, that's a good change with that assy being re-created every time. The lookup is really first part and then Mono.Cecil is adding InternalsVisibleTo to that assemblies so they don't later complain if hot-reloaded code is using 'internal'.

I've still got a task to sort of the other bit so we don't have to recreate those 'helper' assys every time (#33)

How good is overal time improvement (for whole process) on your end when changing same file for first time and subsequent changes?