@SamPruden 2023-06-25
> I've still got a task to sort of the other bit so we don't have to recreate those 'helper' assys every time (https://github.com/handzlikchris/FastScriptReload/issues/33)

I hadn't seen that, but I was also thinking of putting a cache around the whole thing. Maybe I'll get around to a PR for #33. I don't think I was seeing that be very slow after this change, but I should re-check my instrumentation because I could be wrong about that!

> How good is overal time improvement (for whole process) on your end when changing same file for first time and subsequent changes?

I'm still seeing two categories of first run being slow.

 - The first run after a domain reload is horrible. Like, 6-10s! I haven't fully dug into why, yet. Initial profiling seemed to suggest that this cost is distributed pretty evenly across all parts of the code (including Roslyn bits), suggesting a JIT issue or something. But that's a **huge** slowdown I'm not used to seeing in other parts of the Unity code, so it's probably not that. Maybe it's lots of small things caching, including in Roslyn code. I intend to do a proper investigation on this sometime soon.
 - The first run on a file is about 1.5x slower.


Here are some timings running FSR on my 400 line player controller (the file that runs slowest for me) after a domain reload. Edits are only to whitespace. This includes the changes from my other PRs, but they shouldn't substantially change these numbers. N = 1.

Before this PR:
 1. compilation: 6,588ms (of which 455ms in Roslyn process), hot-reload: 39ms
 2. compilation: 2,843ms (of which 509ms in Roslyn process), hot-reload: 18ms
 3. compilation: 2,038ms (of which 389ms in Roslyn process), hot-reload: 18ms
 4. compilation: 1,746ms (of which 368ms in Roslyn process), hot-reload: 17ms
 5. compilation: 1,628ms (of which 341ms in Roslyn process), hot-reload: 20ms

After this PR:
 1. compilation: 5,792ms (of which 480ms in Roslyn process), hot-reload: 27ms
 2. compilation: 513ms (of which 325ms in Roslyn process), hot-reload: 14ms
 3. compilation: 554ms (of which 403ms in Roslyn process), hot-reload: 16ms
 4. compilation: 915ms (of which 478ms in Roslyn process), hot-reload: 18ms
 5. compilation: 504ms (of which 331ms in Roslyn process), hot-reload: 15ms

However, most of that initial cost is due to the first run after the domain reload, not the first run of the file.

 1. (different/smaller file) compilation: 4,883ms (of which 325ms in Roslyn process), hot-reload: 24ms
 2. (first player.cs edit) compilation: 892ms (of which 448ms in Roslyn process), hot-reload: 26ms
 3. compilation: 513ms (of which 318ms in Roslyn process), hot-reload: 13ms
 4. compilation: 619ms (of which 437ms in Roslyn process), hot-reload: 18ms
 5. compilation: 543ms (of which 336ms in Roslyn process), hot-reload: 19ms
 6. compilation: 482ms (of which 332ms in Roslyn process), hot-reload: 21ms

These tests are done on a fairly mid range laptop. They're probably a bit slower than you're used to seeing, but as most of this is single core bound, we probably shouldn't ever be seeing OOM level differences between timings on my machine and yours from hardware alone.

More investigation needed, but we're trending in the right direction. #90 also makes a huge difference to practical speed (shaves off about 2s of filewatcher overhead before FSR triggers).